move collection and data_set
