+++ src/dray/filters/isosurfacing.hpp
@@ -6,8 +6,8 @@
 #ifndef DRAY_ISOSURFACING_HPP
 #define DRAY_ISOSURFACING_HPP
 
-#include <dray/data_set.hpp>
-#include <dray/collection.hpp>
+#include <dray/data_model/data_set.hpp>
+#include <dray/data_model/collection.hpp>
 
 #include <utility>
