+++ src/dray/filters/redistribute.hpp
@@ -1,7 +1,7 @@
 #ifndef DRAY_REDISTRIBUTE_HPP
 #define DRAY_REDISTRIBUTE_HPP
 
-#include <dray/collection.hpp>
+#include <dray/data_model/collection.hpp>
 
 namespace dray
 {
