+++ src/dray/filters/vector_component.hpp
@@ -1,7 +1,7 @@
 #ifndef DRAY_VECTOR_COMPONENT_HPP
 #define DRAY_VECTOR_COMPONENT_HPP
 
-#include <dray/collection.hpp>
+#include <dray/data_model/collection.hpp>
 
 namespace dray
 {
