+++ src/dray/queries/lineout.hpp
@@ -1,7 +1,7 @@
 #ifndef DRAY_LINEOUT_HPP
 #define DRAY_LINEOUT_HPP
 
-#include <dray/collection.hpp>
+#include <dray/data_model/collection.hpp>
 
 namespace dray
 {
